Refactor, AlertRepresentable을 확장 구현하는 방법으로 Alert표출 시스템 개선
