+++ Projects/Feature/DesignSystem/Feature/Sources/Components/Alert/Type2/DSTwoButtonAlertTranstionDelegate.swift
@@ -1,0 +1,60 @@
+//
+//  DSTwoButtonAlertTranstionDelegate.swift
+//  FeatureDesignSystem
+//
+//  Created by choijunios on 1/20/25.
+//
+
+import UIKit
+
+final class DSTwoButtonAlertTranstionDelegate: NSObject, UIViewControllerTransitioningDelegate {
+    
+    func animationController(forDismissed dismissed: UIViewController) -> (any UIViewControllerAnimatedTransitioning)? {
+        CustomAlertDismissAnimation()
+    }
+    
+    func animationController(forPresented presented: UIViewController, presenting: UIViewController, source: UIViewController) -> (any UIViewControllerAnimatedTransitioning)? {
+        CustomAlertPresentAnimation()
+    }
+}
+
+
+// MARK: CustomAlertPresentAnimation
+fileprivate final class CustomAlertPresentAnimation: NSObject, UIViewControllerAnimatedTransitioning {
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0.2
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        guard let toView = transitionContext.view(forKey: .to) else { return }
+        let containerView = transitionContext.containerView
+        
+        toView.alpha = 0
+        containerView.addSubview(toView)
+        
+        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
+            toView.alpha = 1
+        }) { _ in
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
+        }
+    }
+}
+
+
+// MARK: CustomAlertDismissAnimation
+fileprivate final class CustomAlertDismissAnimation: NSObject, UIViewControllerAnimatedTransitioning {
+    func transitionDuration(using transitionContext: UIViewControllerContextTransitioning?) -> TimeInterval {
+        return 0.2
+    }
+    
+    func animateTransition(using transitionContext: UIViewControllerContextTransitioning) {
+        guard let fromView = transitionContext.view(forKey: .from) else { return }
+        
+        UIView.animate(withDuration: transitionDuration(using: transitionContext), animations: {
+            fromView.alpha = 0
+        }) { _ in
+            fromView.removeFromSuperview()
+            transitionContext.completeTransition(!transitionContext.transitionWasCancelled)
+        }
+    }
+}
